ha-mcp/requirements.txt
PyMuPDF>=1.23.0
flask>=3.0.0
anthropic>=0.40.0
ddgs>=7.0.0
mcp[cli]>=1.0.0
python-dotenv>=1.0.0
